requirements.txt
requests
google-api-python-client
pymupdf
python-docx
ebooklib
markdownify
python-dotenv